+++ docs/changes.md
@@ -1,4 +1,4 @@
-# 0.2a0 Release Notes
+# {{ version }} Release Notes
 
 Here you'll find a list of changes in this release, along with a list of what's to come.
 
@@ -8,6 +8,11 @@ Here you'll find a list of changes in this release, along with a list of what's 
 
 Requests are globally rate limited using the Sliding Window algorithm.
 Default setting is 250 requests / 30 seconds, but can be changed via `http.set_ratelimit`.
+
+## Changes
+
+- Added `http.File`
+- Changed FormData functionality
 
 ## To Do